Check if symlinks are valid
When using root option and $initd/$bn is a symlink, the script would fail because
the symlink points to a path on target. For example:

/path/to/target/rootfs/etc/init.d/syslog -> /etc/init.d/syslog.busybox

Hence, [ -f /path/to/target/rootfs/etc/init.d/syslog ] condition would return
false.

This patch adds the posibility to check whether the file the symlink points to
actually exists in rootfs path and then continue.